[Mono.Android] Refactor container factory into abstract base class
Convert ISafeContainerTypeFactory into an abstract SafeContainerTypeFactory
base class and move the MakeGenericType and CreateInstance reflection helpers
(with their trimming/AOT suppressions) onto it as protected members, so the
per-container factories use them via inheritance instead of sharing them
through the entry-point class.

MakeGenericType now takes a concrete reference-argument wrapper exemplar
(e.g. JavaList<IJavaPeerable>) whose DynamicallyAccessedMembers annotation
roots the __Canon template, making the AOT safety obvious. The dishonest
[return: DAM] is dropped in favor of scoping the constructor suppression to
CreateInstance. ValueTypeFactory's mixed reference/value dictionaries adopt
the same exemplar pattern, removing the dedicated rooting-token fields.

Also rename the dictionary GetValueTypeFactoryOrThrow helper to a pure
TryGetPrimitiveValueTypeFactory Try-pattern method, with an up-front
EnsureReferenceOrPrimitive validation, and revert the entry-point class name
back to SafeJavaCollectionFactory to reduce churn.

Copilot-Session: 50ab60e5-5b4d-4e5f-abd2-2878d87a10c9
