+++ src/Mono.Android/Java.Interop/ValueTypeFactory.cs
@@ -2,7 +2,10 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
+using System.Reflection;
 
 using Android.Runtime;
 
@@ -53,20 +56,14 @@ abstract class ValueTypeFactory
 
 	internal abstract IDictionary CreateDictionaryWithReferenceValue (Type valueType, IntPtr handle, JniHandleOwnership transfer);
 
-	internal abstract IDictionary CreateDictionaryWithKey<[DynamicallyAccessedMembers (SafeContainerConverterFactory.Constructors)] TKey> (
+	internal abstract IDictionary CreateDictionaryWithKey<[DynamicallyAccessedMembers (SafeJavaCollectionFactory.Constructors)] TKey> (
 		ValueTypeFactory<TKey> keyFactory,
 		IntPtr handle,
 		JniHandleOwnership transfer);
 }
 
-sealed class ValueTypeFactory<[DynamicallyAccessedMembers (SafeContainerConverterFactory.Constructors)] T> : ValueTypeFactory
+sealed class ValueTypeFactory<[DynamicallyAccessedMembers (SafeJavaCollectionFactory.Constructors)] T> : ValueTypeFactory
 {
-	// These tokens root the mixed reference/value dictionary canonical shapes. For example,
-	// JavaDictionary<string,int> canonicalizes like JavaDictionary<__Canon,int>, so
-	// JavaDictionary<object,T> supplies the NativeAOT template when T is a mapped value type.
-	static readonly Type ReferenceKeyDictionaryType = typeof (JavaDictionary<object, T>);
-	static readonly Type ReferenceValueDictionaryType = typeof (JavaDictionary<T, object>);
-
 	internal ValueTypeFactory ()
 	{
 	}
@@ -97,19 +94,60 @@ sealed class ValueTypeFactory<[DynamicallyAccessedMembers (SafeContainerConverte
 
 	internal override IDictionary CreateDictionaryWithReferenceKey (Type keyType, IntPtr handle, JniHandleOwnership transfer)
 	{
-		_ = ReferenceKeyDictionaryType;
-		var dictionaryType = SafeContainerConverterFactory.MakeGenericType (typeof (JavaDictionary<,>), [keyType, typeof (T)]);
-		return (IDictionary) SafeContainerConverterFactory.CreateInstance (dictionaryType, handle, transfer);
+		// JavaDictionary<keyType, T> canonicalizes like JavaDictionary<__Canon, T>; the
+		// JavaDictionary<IJavaPeerable, T> exemplar roots that template with its constructors.
+		return CreateReferenceMixedDictionary<JavaDictionary<IJavaPeerable, T>> ([keyType, typeof (T)], handle, transfer);
 	}
 
 	internal override IDictionary CreateDictionaryWithReferenceValue (Type valueType, IntPtr handle, JniHandleOwnership transfer)
 	{
-		_ = ReferenceValueDictionaryType;
-		var dictionaryType = SafeContainerConverterFactory.MakeGenericType (typeof (JavaDictionary<,>), [typeof (T), valueType]);
-		return (IDictionary) SafeContainerConverterFactory.CreateInstance (dictionaryType, handle, transfer);
+		// JavaDictionary<T, valueType> canonicalizes like JavaDictionary<T, __Canon>; the
+		// JavaDictionary<T, IJavaPeerable> exemplar roots that template with its constructors.
+		return CreateReferenceMixedDictionary<JavaDictionary<T, IJavaPeerable>> ([typeof (T), valueType], handle, transfer);
 	}
 
-	internal override IDictionary CreateDictionaryWithKey<[DynamicallyAccessedMembers (SafeContainerConverterFactory.Constructors)] TKey> (
+	/// <summary>
+	/// Builds a mixed reference/value <see cref="JavaDictionary{K,V}"/> whose reference argument rides the
+	/// <c>__Canon</c> template rooted by the <typeparamref name="TExemplar"/> instantiation.
+	/// </summary>
+	/// <typeparam name="TExemplar">
+	/// The <c>JavaDictionary&lt;IJavaPeerable, T&gt;</c> / <c>JavaDictionary&lt;T, IJavaPeerable&gt;</c> exemplar.
+	/// Its <see cref="DynamicallyAccessedMembersAttribute"/> roots the constructors of the canonical
+	/// template that <paramref name="arguments"/> resolves to (the exact-value argument stays value-specific).
+	/// </typeparam>
+	/// <param name="arguments">The <c>[key, value]</c> generic arguments, exactly one of which is a reference type.</param>
+	[UnconditionalSuppressMessage ("AOT", "IL3050:RequiresDynamicCode",
+		Justification = "NativeAOT's Type.MakeGenericType() and Activator.CreateInstance() are annotated because arbitrary constructed generics can lack a runtime template. " +
+			"This helper only closes JavaDictionary<,> over one reference argument and the mapped value type T. " +
+			"The reference argument canonicalizes to __Canon and T stays value-specific, so the result shares the JavaDictionary<IJavaPeerable, T> / JavaDictionary<T, IJavaPeerable> " +
+			"template that TExemplar already roots.")]
+	[UnconditionalSuppressMessage ("Trimming", "IL2055:MakeGenericType",
+		Justification = "The generic definition comes from the known JavaDictionary<,> wrapper via TExemplar, not an arbitrary user type. " +
+			"The wrapper constructors are rooted by the DynamicallyAccessedMembers(Constructors) annotation on TExemplar.")]
+	[UnconditionalSuppressMessage ("Trimming", "IL2072:UnrecognizedReflectionPattern",
+		Justification = "The constructed dictionary type rides the TExemplar canonical template whose constructors are rooted by DynamicallyAccessedMembers(Constructors). " +
+			"Only the known JavaDictionary<TKey,TValue> constructor is invoked here.")]
+	static IDictionary CreateReferenceMixedDictionary<[DynamicallyAccessedMembers (SafeJavaCollectionFactory.Constructors)] TExemplar> (
+		Type[] arguments,
+		IntPtr handle,
+		JniHandleOwnership transfer)
+	{
+		Debug.Assert (typeof (TExemplar).IsGenericType && typeof (TExemplar) != typeof (TExemplar).GetGenericTypeDefinition ());
+
+		var dictionaryType = typeof (TExemplar).GetGenericTypeDefinition ().MakeGenericType (arguments);
+		var instance = Activator.CreateInstance (
+			dictionaryType,
+			BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance,
+			binder: null,
+			args: [handle, transfer],
+			culture: CultureInfo.InvariantCulture);
+		if (instance == null) {
+			throw new InvalidOperationException ($"Unable to create an instance of collection type '{dictionaryType}'.");
+		}
+		return (IDictionary) instance;
+	}
+
+	internal override IDictionary CreateDictionaryWithKey<[DynamicallyAccessedMembers (SafeJavaCollectionFactory.Constructors)] TKey> (
 		ValueTypeFactory<TKey> keyFactory,
 		IntPtr handle,
 		JniHandleOwnership transfer)
